+++ Integra/app/src/main/java/com/dhbw/cas/integra/data/TaskDao.kt
@@ -1,7 +1,8 @@
-package com.dhbw.cas.integra.ui.catalogue
+package com.dhbw.cas.integra.data
 
 import androidx.lifecycle.LiveData
 import androidx.room.*
+import com.dhbw.cas.integra.data.Task
 
 @Dao
 interface TaskDao {
